Fixed #255: allow recreation of an order
This allows the recreation of an existing order under certain conditions:
- Orders marked as FAILED and CANCELED can be recreated
- IN_PROGRESS order can only be recreated if in `pending_expiry` substatus
Other IN_PROGRESS orders have no reason to be recreated. Should one want to recreate it
it must be canceled before.

COMPLETED and IN_PROGRESS Orders, when recreated skip the units decreasal
as the user would have already been charged for that order.

Recreation is done via a `re-create` button on the Order detail from the `/all-orders` list.
There's a UI confirmation after click on that button.

Recreation is thus reserbed to cardshop admin and should be used we caution.
There is no ability to customize anything from the Order. Should anything be changed, one
can create a new order.