Fixed #237: cancel tasks marked marked for cancelation
manager allows to cancel manager order.
On manager Order cancel, a request to scheduler is made to cancel order
On scheduler order cancel, all associated tasks's statuses are updated to canceled.

This now checks for the task's status on every poll call to check whether it has been
set to canceled. If it's the case, we stop the task and move on.