Fixed Place an Order selection
Place an Order automatically selects the most appropriate configuration.
This is supposed to be the last modified.
Code was using `.last()` on queryset which fetches the last from the QS
But since the default ordering in the model changed, the QS now returns DESC
configurations based on ID (so latest created first).
Using `.last()` thus results on picking the oldest configuration which is not what
we want.
This change explicitly selects the latest by modification time