Fixed SSID being `default` on auto-images
AutoImages uses a special mechanism in that it exports its config as JSON
at creation/update and is converted to YAML at time of build so it benefits
from all updates.
When we introduced the custom SSID, we did not include it in JSON export
and the JSON parser wasn't looking for it either.

This resulted in autoimages to use the default (`default`) SSID