Fixed error on shop success for wikipedia-en-h1
Success webpage provides feedback on order after Stripe payment.
It is not involved in product delivery but it shows different information
based on the ordered product.
Kind of product is detected from product code.
As we were testing for `wikipedia-*` before testing for suffix `*-h1`,
ordering `wikipedia-en-h1` would match the wikipedia image instead of
the device and attempt to show the image feedback page which failed.