Prevent auto-deletion of autoimages on Wasabi bucket
Our AutoImages have an `expiration` date.
This date used to be set to end of creation month and this date was used
to trigger new Order for the AutoImage.

When switching to manual releases, we simply stopped setting
an expiration date the scheduler would never re-trigger a new Image.

What was overlooked (because it's only in the Creator, which wasn't touched)
is the fact that when we upload to Wasabi, we set an auto-delete date
to 40d after the creation.

There is no such thing as no-expiration for the bucket because we'd
otherwise have to take care of every file deletion. Our bucket is thus
configured for auto-deletion with a default date of 1d.

This commit fixes this issue by periodically checking the actual
autodeletion date on Wasabi for auto-images and should it be considered
soon-enough, it pushes it a little.
This way, once a new image is released, the old one should remain only
for a few days.

CHANGES:

- [health-check] now has existence test for each auto-image. If it's gone,
it fails. We formerly were testing this by accident in another test
and just for the demo image.

- [health-check] now has a deletion test for each auto-image. If any
image is to be deleted within 5days, it fails.

- [periodic-tasks] no checking actual deletion dates for all ready autoimages.
If soon-enough (5d), it extends it (+10d).
Can be tweak via AUTO_IMAGES_EXTEND_BEFORE_DAYS and AUTO_IMAGES_EXTEND_FOR_DAYS

- [scheduler] Indiv `/auto-images/:slug` endpoint now includes an `autodelete_on` (ISO formatted)

Fixes #483