Fixed multiple config creation instead of update
When creating a new configuration, user would be creating a new
config for every hit to the save button was considered as the first one.

Instead of fumbling with mechanism, this ensures that first time
saving redirects to edition view instead.
Allows user to save/share URL from first save.

Fixes #496